backend/grocery_assistant/requirements.txt
django
python-dotenv
psycopg2
flake8
djoser
djangorestframework
django-colorfield
Pillow
django-allauth